// src/types/PostTypes.ts
export interface PostRateType {
  Cur_ID: number;
  Cur_Abbreviation: string;
  Cur_Scale: string;
  Cur_Name: string;
  Cur_OfficialRate: string;
  error?: string;
}

export interface PostCurrencies {
      Cur_ID: number,
      Cur_Code: string,
      Cur_Name: string,
      Cur_Name_Eng: string,
      Cur_Scale: string,
 }